Maybe solved the object/record/table issue
I need to write more tests, but it looks good so far. The NamType record
can be extended to provide whatever information I need down in the
translators. The record/table/object name can be used to specialize the
construtor type.

C++ still has a bit of a problem. Since there is no hetlist type in C++
(well, tuples could be used, I guess), I generate a structure that
stores the required information. This works fine, except that the
generated struct is not in the scope of the C++ code where the
programmer wrote it. This also is fine, but the programmer would have to
write the functions using the records as templates. For example:

``` C++
template <class R>
int yearsLeft(R person){
    int years = 80 - person.age;
    return years;
}
```

This is should be fine, though. If a C++ function is specified as being
a record, then a type will automatically be defined for it. If the user
wants to use an existing record, then they can use the object keyword
instead of the record keyword.