(7/304 fail) Add table/object tests
All the C++ that uses records is now broken since that last commit. The
reason is that I replaced the record names with the constructors
everywhere.

Now I write this:

```
record Cpp (Person a) = "struct" {person :: "std::string", info :: a}
```

Instead of:

```
record Cpp (Person a) = "person" {person :: "std::string", info :: a}
```

Here the "person" is the name of the generated struct in C++. In R and
python, I use the constructors "list" and "dict", respectively. In R and
python these types are heterogenous lists. They do not, by themselves,
represent the full type, thus they are the dynamic homologs of C
structs. However, unlike "list" and "dict", the C structs require a
name. Also, back when I named the record types (e.g., "person"), I
unified all "person"s across all the types they might carry and generate
templates accordingly. For example, it would unify (Person Int) and
(Person Str) to make a struct template of one generic parameter. Now
that all automatically generated records have the same name of "struct",
I need some new term to unify around. The most obvious choice is a
name-mangled version of the general type name "Person", for example
"mlc_Person". But "Person" is stored as the type alias which currently
disappears in the frontend desugar step. I could put the "Person" into
either the NamT type or the NamType type.

I could also not store the "Person" anywhere in the concrete type object
and instead pass it over be comparing the type to the general type
further down the line. That would require more wiring, I think for now I
will just pack it into the NamType and refactor later if I so fancy.