(9/304 fail) Refactor with TypeP
Now all types after realization are based on TypeP, which is exactly
Type with TVar replaced with PVar. Where TVar stores the language and a
type variable but PVar additionally stores the general type variable.
So TypeP merges the general type and a single concrete type. This makes
the general type name for records available to the translators (which
was the initial goal of the refactor) and also couples concrete record
keys to general keys (which will be important in the near future).

Currently every C++ program with records is failing since the concrete
name for all records is "struct". For these automatically generated
structs, I need to replace the name with some mangled form of the
general name.