(5/307 fail) All anonymous records work (fix #8)
I guess this is later (see previous commit). I worry that my C++
solution is fragile since it is dependent on prototype declaration
order. It definitely would fail on a recursive struct definition, but
this shouldn't be possible for anonymous types. Also, the anonymous
structs should probably be using pointers if they wrap another struct.

Meh, I'll fix it later ...