Pass GMeta all the way to RecEntry ...
To make C++ work with parameterized records, I need to know the types in
the parameters, however this information currently disappears in the
desugaring step. There is a difference in the way parameterized types
(ArrT) and parameterized type defs work, the later is kept all the way
forward, the former disappears. I could add a parameter list to NamT and
typecheck it just as I typecheck ArrT.