(2/307 fail) All R tests pass
I am not pleased with the direction serialization is going. All three
languages use different approaches for the job. In C++, I generate
serialization code with type schema used to resolve templates, in R I
have a dedicated C++ wrapper that parses a JSON file describing the
type, and in Python I have a wonky type description with constructor
functions nested inside.

The last failing tests are for template classes in C++. I need to know
the types of the record parameters in the C++ constructors, but this
information is lost in the frontend desugar step. Lack-a-day, what can I
do?