(3/315 fail) Basic code done for record access
See the three failing tests. There is a problem in the parser, since
`sum x@a` does not parse correctly. The expression `sum x` is
interpreted first as an App and then the parser tries to access from the
record `sum x`. I need to ensure that `@` binds more tightly than
function application. So, I need to add logic for binary operators.

For now, I can just wrap record accesses in parentheses.