All tests pass
I had to change the parser to return non-existential default container
types. That is, in a concrete signature, if the forms `[a]` or `(a,b)`
are now typed to the first default value rather than to an existential
type. If the user wants to use an alternative list or tuple type, they
can do so explicitly (e.g., `std::array<$1> double`).