A partial solution to the primitive problem
For now I can add primitive terms to the DefaultTypes lists. Then, if
the primitive is covered in the language serializer, everything is fine.
I added `long` and `size_t` types to C++, which solves the immediate
problem. The function that creates the SerialAST trees checks these
defaults to determine the JSON type of a type term. However, this
delegates the handling of primitives to the morloc compiler, when I
would prefer it be implemented in libraries.

The full solution to this problem will probably involve typeclasses,
which I am deferring to a future dev cycle.