Replace `lib` with `src` in MORLOCHOME
Previously I automatically installed morloc modules in a
`$MORLOCHOME/lib` folder. However, this is source code, so, following
UNIX tradition, should go in a `src` folder. `lib` should be reserved
for shared libraries. When a C++ program is installed, a shared library
should be made and moved to the `lib` folder. The header file should be
moved to an `include` folder. When the C++ source is imported, the
header file should be read and the shared library linked.

The C++ import name can be appended with ".h" to make a header file and
converted to lower case, prepended with "lib", appended with ".so", and
have all non-alphanumeric characters removed to get a shared library
name.

Given the import command

`source Cpp from "SimplexNoise" ("SimplexNoise::noise" as noise)`

First a header would be sought in the following locations:
    * ./SimplexNoise.h
    * ./include/SimplexNoise.h
    * ./src/SimplexNoise.h
    * $MORLOCHOME/include/SimplexNoise.h
    * /usr/include/SimplexNoise.h

If no header is found, raise an error, otherwise:

Search for a shared library in the following locations:
    * ./libsimplexnoise.so
    * ./bin/libsimplexnoise.so
    * ./src/libsimplexnoise.so
    * $MORLOCHOME/lib/libsimplexnoise.so
    * /usr/lib/libsimplexnoise.so

If an library is found, use it, otherwise look in the folder containing
the header and try to build a shared library there. On success, move the
library to `$MORLOCHOME/lib/libsimplexnoise.so`.

Then the whole thing would be built with a command such as:

```
g++ pool.cpp src/SimplexNoise.h -I$PWD/src -Wl,-rpath=lib -L$PWD/lib  -lsimplexnoise
```