Update demo writeFasta command
Now it actually makes a file rather than just returning a FASTA string.

Should a write function return "void" or the path to the file that was
written (if successful)? The former is certainly the more common
convention in C++ (and other languages). And it is simpler and makes for
more obvious type signatures. The latter, however, makes pipeline
slightly easier if the filenameis needed downstream (which it often
would be in the bioinformatics setting).

Though actually the reason I returned a string was because my void
handling for C++ is crap.