Wire the new command line argument parser
It looks beautiful so far, I still need to test it thoroughly

Next I need to wire the `-o,--outfile` option so that `nexus.pl` can be
replaced with a custom name.

Along those lines, I really also should build the program in a
MORLOC_HOME folder and make the nexus use absolute filepaths (so that it
can be moved anywhere on the filesystem). Perhaps, though, this should
be done instead by the `morloc install` command with no arguments OR by
`morloc make --install`. This would move the executable to somewhere on
path and the pools to a stable location in MORLOC_HOME.