Update changelog
The only thing left to do in this dev cycle is to replace the "nexus.pl"
hard-coded executable name with an option via the `-o` flag.

I need this in order for the examples in my poster to run. Well, at
least it will make them closer to running, the poster still has the
braces around modules.

I would also like to get really pretty error messages working. There is
no reason for me to procrastinate it any longer, especially since I am
beginning to understand how to do it. Probably it would take only a day
to get something acceptable together. The current walls of inscrutible
text are a hard turn-off to anyone experimenting with the language.

The focus of cycle v0.35.0 is to get Rust support on-line. There are a
few issues we'll need to sort out for C++ and Rust to work well with
morloc. I don't think we'll solve all of them in this cycle, but we can
identity them and outline rough solutions.