Delegate more to the generators
Now the Script record stores a directory of generated code
(`AnchoredDirTree`) instead of just one file. I also remove the
`scriptCompilerFlags` and `scriptInclude` fields, replacing them with a
single `scriptMake::[SysCommand]` field. Now the record stores a
directory of generated code which will allow more idiomatic package
generation. The `scriptMake` package stores a list of generic operations
that should be performed after writing the directory.

The immediate motivation for this change was the need to generate more
structured Rust architectures, for example:

```
<package>
  |- Cargo.toml
  '- src
     '-main.rs
```

Where the Cargo.toml file stores (among other things) the dependencies
the `main.rs` file relies on.

`morloc` should be able to source directories of source code that have
their own build scripts.

TODO: replace `scriptBase` with `scriptExec`. Currently the pool
generators and the nexus generator are coupled awkwardly through the
`Morloc.Language` function `makeExecutableName`. A much smarter approach
would be for the pool generators to run first and return the Script
records as instructions for building the pools AND with links to the
executable. The latter data will be passed to the nexus and where these
executables will be called as needed.