Add basic namespace handling for Rust
Still need thorough tests. But the basic idea is this, find the
dependency name for base, collapse with "::", and voila, magic.

So:

```rut
use foo

foo::function
```

`foo` is the namespace I add and this namespace I get from the package
name, it is the root (e.g., "crate::foo"). But I don't do the crate
thing.

Now, if I need need access subspaces within the crate, then, in morloc
space, do something like this:

```morloc
source Rust from "." ("bar::thingy" as thingy)
```

This would compile to `foo::bar::thingy`, since the root name will be
collapsed with the source name.

The last thing breaking in `argument-form-1-rs` is the `String` return
type of the manifolds. The immediate solution would be to separate the
input from return serial types into "&str" and "String", respectively. I
shall have to consort with the jenra about performance before coming to
a conclusion.