(1/316) Fixed parser error
In the expression

```
f3 = x where
    x = 42
```

A syntax error was being raised. The reason was that I was missing a
`try`. Simple.

Now the same expression fails with an `UnboundVariable: x` error. This
indicates that the scoped `x` variable could not be found in `Infer.hs`.
So, fix this next, but I have to run to work.