Fix pretty printing of EVar to include scope
Terms defined in `where` statements are not being found when used in the
body of a declaration because they have not been declared. By the time
Infer.hs sees the code, the `where` statements have been resolved into a
linear series of declarations of terms with non-empty scopes. The
typechecker loads these into the context Gamma in order and, currently,
terms must be defined before they are used. Now, the order in which
terms are defined in morloc *should* not matter. This order dependence
is an old bug -- perhaps the time has come to fix it.