(4/322) Paint little flowers on it
The current scoping scheme breaks when there are multiple
implementations of a term. For example:

```
inc :: Num -> Num
inc py :: "float" -> "float"

add :: Num -> Num -> Num
add c :: "double" -> "double" -> "double"

addTwo x = inc (inc x)
addTwo x = add x 2
```

This is a perfectly reasonable thing to write. But the scoping sets the
x's in the two implementations to `addTwo/x`. In the first
implementation, `addTwo/x` is assigned the type of `float@py` and `Num`. Then
in the second implementation, `addTwo/x` is looked up and these old
types are found and upon checking the `double@c` is added. The two `x`s
should be distinct, so I need a way to separate them.

I need to convert the data into SAnno before typechecking. This will
remove the morloc composition abstractions and module abstractions. Then
I can infer which nodes in the SAnno tree are in which languages
(basically by assigning existential types down the tree).

The SAnno trees naturally handle the multiple-implementations problems
and avoid some of the lookup and ordering logistics of my current setup.
Also, the where statements and scoping issues vanish and I can remove
the paths from the EVar again.