Add Rose tree module
Handling scope by adding paths to the VarE type, which stores term
names, was a bad idea. Since there may be multiple implementations of a
term, paths may be ambiguous.

Instead, scoping can be handled using a normal old Rose tree.

```
data Rose = Nil | Rose a [a]
```

The parse can build a `Rose Expr` tree. I need to add a `TopE`
constructor to the Expr type that represents the root of a module and
the `Nil` of the Rose.

To extend this to the Parser, I need to convert ParserNode from

```
data ParserNode = ParserNode  {
    parserNodePath :: Maybe Path
  , parserNodeBody :: [Expr]
  , parserNodeSourceMap :: Map (EVar, Lang) Source
  , parserNodeTypedefs :: Map TVar (UnresolvedType, [TVar])
  , parserNodeExports :: Set EVar
}
```

to just `Rose Expr`.

The parserNodePath data is now stored as the first TopE element in the
Expr constructor TopE. The constructors parserNodeSourceMap,
parserNodeTypedefs, and parserNodeExports are all used only once and
may be replaced with function calls on site.