Desugar code passes the typechecks
I need to walk through the code very carefully to ensure I know exactly
what is happening and then I need to concisely document it.

Also the packer map handling is currently all commented out. It needs to
be moved out of the desugar module and into the treeify module.
Previously, packer information was rolled into the PreparedNode objects.
Those no longer exist, so the info must be stored in the DAG/Expr until
such time as they can be accounted for in the GMeta objects. Although,
storing it in gmeta is also far from ideal, since this leads to massive
duplication.