Add a non-injective, non-surjective, single value copy map (GMap)
`GMap` has the type `GMap (Map a b) (Map b c)`.

The first map maps to keys of the second map. The initial use case is to
store links to a type annotations shared by many terms. This reduces
duplication in memory and allows easy unification of terms. For example,

```
x :: A

foo y = baz (bar x) x

bar z = z + 1
```

Here both `x` terms in the definition of `foo` and the `z` in `bar` are
all linked to the type `A`. When we see that `x` is passed to `bar`, we
can set the type annotation of `z` to the same as `x`.

Or we can take a SAnno object and reindex it by creating new indexes and
linking them to whatever the old indices pointed to.