Indexing compiles
All data from type signatures is now written into a symbol table stored
in MorlocMonad state and indexed with the ExprI indices mapping them to
terms in the input code.

It is still pretty messy and I'm not sure the logic is right, but is
passes the typechecker.

Next I need to implement `collect`, which should be easy since all the
lookup mess is now gone.

Then there is the typechecking, which will need to ... be ... rewritten
(sob sob)