(78 undefined) it passes the typechecker
Am I misguided in my choice to through so much into state? The approach
avoids the need to manuall pass a giant table all through my program.
However it also means that I have to do more in the monad.

The codebase is in dire need to code cleaning and documentation.