Partial implementation of realize
I can realize gefore segmentation. This gives a concrete type of

 `SAnno (Indexed Int) One (Indexed Lang)`

This type may be immediately segmented into pools and then concrete
types can be inferred. This path does not allow type conversion
optimizations. Also, all let-optimizations are removed for now since
they never worked in the first place. The problem with let-optimizations
is that they assume the functions are all pure, which is not valid until
I add an effect system where purity may be specified.

Eight functions remain undefined in Generate.hs. The next thing to do is
complete these and tie everything back into generators further back. I
will also need to figure out packers and other things that used to be
passed in the GMeta objects.

After this I need to jump back to the frontend general type checker.