Define the skeleton of Generate.hs
The basic type structure of Generate.hs is now defined. There are 5
undefined functions that need to be filled in.

Four of these are pretty easy and require a bit of remapping with
indexed types:

 [ ] generalSerial
 [ ] parameterize
 [ ] express
 [ ] getSrcs

One of these is complicated -- the typechecker:

```
typecheck
  :: SAnno (Indexed Type) One (Indexed Lang)
  -> MorlocMonad (SAnno Int One (Indexed TypeP))
```

Contrast this to the general typechecker in Infer.hs:

```
typecheckGeneral
  :: SAnno Int Many Int
  -> MorlocMonad (SAnno (Indexed Type) Many Int)
```

The typechecker in Generate.hs has an unambiguous tree and is dealing
with concrete types. It will also need to check the existence of proper
serialization functions.

There will be some similarity in the bidirectional algorithm both use,
but perhaps not enough to merit generalizing the algorithm to deal with
both ambiguous and unambiguous trees.

Anyway, more on this later. For now I need to focus on implementing the
four "easy" functions. `getSrcs` will actually require finding all
packers that are used, this will require that I go back to the
frontend and capture packer info. So it may be awhile yet before I get
back to implementing this typechecker.