Refactor typechecking code
The code is a lot cleaner if we alternate between SAnno and SExpr (of
course). Synth is now almost completely defined. The last bit I need to
write is definitions for LamS and AppS.

Also the subtype and instantiate functions should not be shared between
the front and back ends because only one of the backend requires
language support, so the frontend can be simplified.