Curry, curry, curry
My typechecking implementation is complicated by my earlier decision to
uncurry my lambda and application expressions. So rather than thinking
about `App Expr Expr` I have to think about `App Expr [Expr]`. Dealing
with this incurs unnecessary bookkeeping and obscures the relationship
to formal descriptions of type system.

However, this general problem is not limited to App and Lam. From the
very beginning I have been separating containers into lists, tuples and
records. Typechecking these is complicated and every operation on types
or expressions must handle all three cases and map all operations over
these lists. These three types can be merged into a Cat type that stores
an annotation (List, Tuple, or Record) and a pair of expressions.

Here are a few examples (for convenience, List=L, Tuple=T, and Record=R
and primitive expressions are left as literals):

 []  -->  Cat L Uni
 ()  -->  Cat T Uni
 {}  -->  Cat R Uni
 [1,2,3]  -->  (Cat L 1 (Cat L 2 3))
 (1,2,3)  -->  (Cat T 1 (Cat T 2 3))
 ((1,2),3) --> (Cat T (Cat T 1 2) (Cat T 3 Uni))
 {a=(1,2), b=3} --> (Cat R (Cat R "a" (Cat T 1 2)) (Cat R "b" 3))

The only difference between the three containers, in general, is the
constraints. The only constraint on the tuple is that it is ordered. The
list is ordered and each left hand expression must have the same form.
Records are unordered and each left-hand expression must be a string.
There are many additional container types that could be added later,
such as sets, and doing so would require only adding a new constructor
to the CatTag data type and implementing container-specific handling
where specifically necessary.

Another advantage of currying every expression is that every element in
a container will be uniquely indexed. This makes distinguishing between
a thing as an element and the thing by itself, which useful since they
may be in different languages.