Despair of ever seeing the light of day
OK, `AppE Expr [Expr]` and other wide style types are nifty for all the
reasons described in the prior post. But here are some problems.

Degenerate representations exist for all the wide types. For example,
`AppE f [1,2,3]` can be written as `AppE (AppE f [1]) [2,3]`. I can
collapse these, but that is an extra step and a extra edge case to
always check for.

This may not be a damning problem.