Restore missing parameters in Nam* types
In the previous refactoring I had removed the name and metavariable
parameters from the record/object/table types. Also the naming
convention `Nam*` should be preferred to `Rec*` since they represent
more than just records. More generally, they are unordered structures
with names. They may be records. Or they may be tables (where each key
is a column name and each value is the type of the elements in the
column). Or they may be inputs to an object constructor.

The whole record/table/object system (let's call it TOR) system is in
need of a heavy update and will be one of the focuese for next year.