(45 undefined) Typechecks with new packer handling
Now packer maps are scoped. They are stored in a `GMap Int MVar PackMap`
structure. This allows every module to have its own set of packers. This
is still not good enough, since I should allow every scope, including
where scopes, to have their own pack handling. But I will worry about
that later (it isn't a hard thing to fix, I just need to generate an
index for the scope and substitute this for the MVar key).

The packer handling as stands is almost certainly wrong. Packers should
be imported implicitly ONLY if the term the (un)pack is imported. For
example, the (un)packers for `Map` will be imported if `Map` is
imported. This issue won't cause many problem, but it is sloppy. It
would only cause problems when conflicting (un)packers are used.

Another issue is that packers are currently not properly typechecked.
Packer handling is not trivial and requires as much care as the more
conventional typechecking. I need to VERY carefully outline the
algorithm that is being followed and thoroughly test the implementation.