Write the general SAnno ExprS typechecking cycle
It is simple enough. SAnno wraps several ExprS which should all have the
same general type. So in synthesis mode, we synthesize the first element
and then check against the SAnno sans the first element.

If in checking mode, we skip the synthesis step and just check all the
way down.