(48 undefined) Wrote concrete App=> rule
It isn't pretty. It probably isn't correct. But it compiles, so there's
that. I need to very carefully go over the `App=>` rule. It seems I
don't exactly need the application step (the 3rd "direction" in
Dunfield's "tridirectional" algorithm).

I'm going to need to do some serious testing of this system.