(40 undefined) Write NamU subtype case
And I must say the new version is a great improvement. The old approach
was to sort the records and then iteratively compare every key/value
pair. Now instead I take the first field/value pair from the first
record and look it up in the second record. If the entry is found, I
remove it from the second record and then subtype the values and the
decremented records.