(38 undefined) Check concrete/general agreement
The concrete type should have the same form as the previously
typechecked general type. I need to give the matter more formal thought,
for now I am requiring the concrete type be a subtype of the general
type. Subtyping across languages is pretty lax, terms needn't have
the same names, for example.

There is also the possibility of inferring concrete types from the
general type under some circumstances. For example, if the general type
is an integer, and nothing can be inferred about the type of the
concrete expression, then a sensible default could be introduced. But
this would only be a guess and guessing is frowned upon in this field.