(2/57 fail) Treeify appears to be broken
Maybe I should create a set of tests for treeify that cover the
translation of ExprI into SAnno. For example,

```
f x = x
f 42
```

Nothing is explicitly exported, so just the one main expression (`f 42`)
is returned as a SAnno. `f` should be expanded here into the lambda. So

```
SAnno (Many [(
  AppS (SAnno (Many [(LamS [x] (SAnno (Many [VarS "x"]) _))], _) _)
  (SAnno (Many [(NumS 42, _)]) _), _)]) _
```