(5/71 fail) How should undefined terms be handled?
Say we have an expression such as:

```
f :: Int -> Bool
f 42
```

This should pass the typechecker. The expression `f 42` has type `Bool`.
Of course, `f` is also undefined. The type is given but no source or
declaration.

How should this be expressed as a SAnno?

```
SAnno (Many []) i
```

This would be fine on the general side. But what if there are language
specific annotations:

```
f :: Num -> Bool
f R :: "numberic" -> "logical"
f 42
```

Now `SAnno (Many []) i` doesn't work.

Further, empty lists sort of catch the idea that there is no expression
for the general type, but it breaks when we realize to `SAnno One`.

We need something well formed to substitute in here. We need to infer a
default expression from a type.  Well, that's some crazy sorcery. Can it
be done? Well, no. Not really.  It would eventually jam something down
in the generator's gears. But, just for fun, let's work through this:

```
FunU ts t = LamS (map make ts) (make t)
LstU ts = LstS (...)
VarU t = ?
```

Well, that didn't work, dreadful idea anyway. Hmm, maybe I need a new
term in SExpr. Maybe `UndefinedS`. This would pass the typecheckers, if
correctly annotated, but would die in the generators. In fact, there
should probably also be a morloc keyworkd `undefined` that is equivalent
to the Haskell `undefined`.