(4/113 fail) More tests, found new error
This snippet fails to fail

```
f :: Num -> Bool
f x = 9999
export f
```

The signature requires a boolean is returned but the declaration returns
an number. The typechecker infers the type `forall a . a -> Num`. It
ignores the annotation when an implementation is available. What it
should do is synthesize a type from the declaration and then check it
against the known implementation.

Actually, for all synth steps, the typechecker should first lookup any
annotations the term has. If any are found, then the synthesized type
should be checked against the annotation.