(12/162 fail) The parser is broken
Types are consuming terms beyond where they should be allowed to.
Whitespace handling in morloc is incorrect and needs to be totally
overhauled.

Here is an example:

```
module Foo
    add :: Num -> Num -> Num
    x = add 1 2
    export x
```

This fails, stating that `=` was found when a toplevel expression was
expected. The reason it fails here is because the last `Num` ate `x` as
a parameters.

An unrelated problem is that the inferred `add` the function type is
actually  `Num -> Num -> Num a`. This implies that Num has both the
kinds `*` and `* -> *`, which should be illegal. This is another species
of bug that the compiler should (but currently doesn't) catch.