(3/175 fail) Replace qualifiers with fresh terms
This fixes the `head(head [[1]])` test case. Here the `a` terms in the
first and second head functions (`[a] -> a`) overwrote eachother,
leading to a subtype operation of `Num <: [Num]`.

I am slightly disturbed, though, because in the past I did not have to
do this substitution. I fear that I lack finesse. I should cut the
context such that these conflicts cannot occur.

Perhaps. Or perhaps my old code was broken as well?