(1/207 fail) Tuple subtyping is not to blame
The `map fst` case is failing in `(Num,Bool) <: (<v10>, <v11>)`. More
specifically, it is failing at the instantiation step `Num <:= <v10>`.

Why? The instantiation function is not the problem. The reason it is
failing is because <v10> is not defined in state. There should be an
`ExistG "v10" [] []` index in Gamma. So to solve this problem, I just
need to find a place where I create an existential but do not store it.