(6/214 fail) Make (App* :: App* Type* [Type*])
Now App is the mirror image of Fun. More importantly, now I can properly
generalize the first term in a parameterized type, so can express neat
things like `forall f a . f a -> a`. It is a bit ugly, since I don't
really support anything other than a VarU as the initial type. But that
can be easily changed when I have a convincing use case to pursue.