(6/214 fail) Better names for existentials
The typechecker debug logs where a mess of v1's and q9's. To ease
debugging, I've made the existential names include user provided names.

To decouple the test suite from the particulars of naming, I rewrite the
existential names in the final test comparison. So in the unit tests,
a return type such as `(<f4> -> <f3>) -> <f3>` would be rewritten to
`(<e0> -> <e1>) -> <e0>`.