(2/213 fail) Remember I do not have sum types
Well, I feel a bit stupid. The test case:

```
out :: f a -> a
out (G 1)
```

Obviously should NOT have type 1. I was subconsciously making the
connection that `G` was a constructor for a parameterized type. But
morloc doesn't know about any of that (and even Haskell wouldn't since I
haven't defined that constructor). All morloc sees is a function named G
that creates something of type `f a`, where `a` could be anything, it
needn't be Num. So I need to make a constructor explicitly.

After adding constructors, all the tests now pass except the mapping
functions that are still failing for the old old reason.

So, need to track that down.