(0/213 fail) Apprehensive happiness
I solved the final problem by deviating from Dunfield's holy word. I
removed the cut from the ForallR subtyping rule. It was cutting away
terms that I needed later. Now, this is a bit concerning, because that
subtyping rule used to work. It is possible removing it, and rewriting
qualifiers in advance to force uniqueness, is masking a downstream
error.

But my tests pass. So happiness?