(1/214 fail) It begins
This is the first successful failure of the backend typechecker. The
wiring for writing tests works. It will require some change since now I
will be checking against TypeP. There is no polymorphism in the backend.
I am tempted to throw out all my old language-specific tests and build
them back up ab initio.

The problem with this first failure is in Treeify. It appears that
concrete types are being written into the first field of TermTypes that
should contain only general types.