(29 undefined) Do repeat yourself
I shamelessly copied the entire typechecker from the frontend back to
the backend.

There are reasons for separating the two. Each one individually is
clearer on its own. The frontend typechecker needn't consider language.
The backend doesn't need to lookup types, since any type it needs
will be annotated at a CallS leaf. The backend typechecker is fully
realized, so there is not need to iterate over expressions. Then again,
the wrapper is a functor so that can be done generally. The backend
needs to check against the inferred general types. It also must deal
with the interfaces between languages and the suitability of the given
packers. The front and back ends may be similar now, but they will
specialize over time.