(3/218 fail) Stubify metaTypedefs
This avoids the undefined error, but will break the C++ translator when
it needs to deal with templates later on.

But I'll deal with that later.

Source code can now be generated in all three supported languages. There
are two apparent errors in the generated code.

 1. The paths to modules are wrong. In python the import:

     `from core import *`

    should be

     `from pybase.core import *`

 2. The manifold indexing is wrong. The pools and the nexus disagree on
    the index of the top manifold.

I can manually fix these issues and recover functionallity of the Python
golden test (argument-form-1-py).