(3/218 fail) Fixed the environment error
I was simply failing to load the module path into parser state. Now to
fix the manifold id error.

In the morloc code

```
import pybase (add, mul)

export foo

foo x = add x (mul 2 20)
```

The nexus is using the id `2` for the toplevel manifold. `2` is the
index associated with the exported `foo`. In contrast, the pool
generator is associating `9` with the same thing. `9` is the index of
the `add` application.