(1/218 fail) Sync nexus and pool indices
Syncing the indices required moving the index from the lambda to inner
body in express'. This overwrites the internal index which could lead to
pain downstream. I need to find a better solution, but it works for now.
`express` is downstream of all typechecking.

C++ and python pass now. R is failing simply because I do not have the
rmorlocinternals package installed. Fixing this means I need the
internet, so I'll need to leave the black couches and visit a cafe.