(91/333 fail) Set 0 default for empty list scores
Stuff is failing on the undefined Lambdas in the translators. This is a
little odd since these cases where not defined in my previously passing
state. The lambdas were rewritten as applications of manifolds. Only
explicit lambdas should be represented as LamM. Something is off in my
wiring, probably in Generate.hs.