(71/333 fail) Fix infinite loop
The bug was a single apostrophy.

`langs' = langs' <> ...`

Where it should have been

`langs' = langs <> ...`

I've also worked out how to do profiling in stack. I know, I'm terribly
late to the game. But in the past I haven't needed it. But these
infinite loop bugs are hard to find without a trace.

I put the following functions in my bashrc:

```
function pinstall (){
    stack build --fast --profile --ghc-options="-fprof-auto-top -Wall" && stack install
}

function prun (){
    stack exec --profile -- $@ +RTS -p -hc -xc
    stack exec -- hp2ps -e8in -c *.hp
}
```

This profiling sort of does everything. Benchmarking and tracing and
even makes a memory usage plot (apparently most of my memory falls into
a pitch black Other hole).

The prun function creates a lot of `morloc.*` debugging files which I've
added to the toplevel gitignore file.

At some point in the VERY distant future, I may try to improve the
performance of the morloc compiler. It is not the bottleneck in building
since compiling the targets is so much slower currently.