(??/fail) Fixed a bug in realize
Not all nodes in the AST were being assigned languages. This can occur
when the top-level expression is a container. Languages are assigned in
a tip to root fashion which guarantees a language bubbles to the top if
any term is concrete inside. But to label the unknown branches, a second
propagation from root to tip is required.