Add 9 scoping tests covering 'where'
Overall, it looks like most of the problems that break the tests are
subtle but trivial logic bugs. The one non-trivial piece of missing
logic is the backend typechecker where I need to ensure the language
interfaces are correct. I should also typecheck the packers.

For now, I will ignore the polyglot cases and ensure everything else is
working correctly. Here is what needs to be done:

 * fix the backend compiler
 * write a systematic tests covering base polyglot cases
 * sweep out the linter warnings
 * removing commented functions from every file
 * build haddocks and check all the documentation
 * push to master