Fix scoping-8 issue and sniff something foul
The scoping-8 tests failed because the index that should have linked to
the packmaps in scope was not in the dictionary. I claimed with great
conviction in the comments that this should not have been possible. I
believe the problem appeared when the first `x` was replaced with an
expression from a nested scope. I am not sure. I need to review the code
and write tests that cover mixing of scopes more rigorously.