(20/351 fail) Fix the serialization
The serialization engine in morloc is one of the newer ideas in the more
intricate pieces of the system. I need to very carefully explain and
document it. It pains me to say that I have no memory of how it works.
The UnresolvedPacker type in particular needs to be much more thoroughly
documented. Serialization was broken because I listed the
unresolvedPackerType as the contracted type (e.g., Map a b) instead of
the expanded type (e.g., ([a], [b])).

Also the boilerplate in morloc is totally out of control. I need to
reimplement with recursion schemes. Well, that can wait until later.