(2/227 fail) Optimize the right thing
I feel pretty ding dang dong stupid. Instead of minimizing cost of
realization, I was maximizing it. Meaning I selected the slowest
possible implementation.

I switched to minimization, but the problem with minimization is that
there is not an obvious default value for `minimumDef`. In a
maximization problem, over positive numbers, I can use 0. I may need to
transform the scores or maybeify them. The only case where the default
score is needed is when no language specfic info is available for an
expression. This case is common since it applies to all primitives and
containers of primitives. These cannot be scored by the bottom up
algorithm, but are rather assigned scores in the later top down
algorithm.

For now, I just inserted 999999999 as the default. Stupid, I know, but
I'm in a hurry.